Move folder black list to watcher module
